Charlatan

+++ src/cogs/admin.py
@@ -17,7 +17,8 @@ class Admin(commands.Cog):
         description="sends a message to the given channel. " +
         "reacting with the given emoji will assign the given role")
     @discord.commands.default_permissions(manage_guild=True)
-    @discord.commands.option('channel', type=discord.TextChannel, parameter_name="channel", required=False)
+    @discord.commands.option('channel', type=discord.TextChannel,
+                             parameter_name="channel", required=False)
     @discord.commands.option("emoji", required=False)
     @discord.commands.option("role", type=discord.Role, required=False)
     async def greeting_role(
@@ -34,14 +35,17 @@ class Admin(commands.Cog):
             try:
                 true_emoji = Emoji(emoji)
             except ValueError:
-                await ctx.respond(f"Whoops! {Emotes.WTF} That emoji is not a valid discord emoji", ephemeral=True)
+                await ctx.respond(
+                    f"Whoops! {Emotes.WTF} That emoji is not a valid discord emoji", ephemeral=True
+                )
                 return
 
         text = text.replace("<<nl>>", "\n")
         try:
             message = await channel.send(text)
         except discord.errors.Forbidden:
-            logger.info("Permission failure for chain_message", guild_id=ctx.guild_id, channel_id=channel.id)
+            logger.info("Permission failure for chain_message",
+                        guild_id=ctx.guild_id, channel_id=channel.id)
             await ctx.respond(
                 f"Whoops! {Emotes.WTF} I don't have permissions to write in {channel.mention}",
                 ephemeral=True
@@ -61,16 +65,15 @@ class Admin(commands.Cog):
             )
         await ctx.respond(f"Message Sent! {Emotes.HEART}")
 
-    @discord.slash_command(name="remove_single_role_assignment", description="Takes out all of Nix's role " +
+    @discord.slash_command(name="remove_single_role_assignment",
+                           description="Takes out all of Nix's role " +
                            "assigning behaviour for this role")
-    @discord.commands.option("role", type=discord.Role, description="The role to remove assignment for")
+    @discord.commands.option("role", type=discord.Role,
+                             description="The role to remove assignment for")
     async def remove_single_role(self, ctx: discord.ApplicationContext, role: discord.Role) -> None:
-        db.multi_void_SQL([
-            ("DELETE FROM RoleChannel WHERE GuildID=%s AND RoleID=%s",
-             (ctx.guild_id, role.id)),
-            ("DELETE FROM ReactMessages WHERE GuildID=%s AND RoleID=%s",
-             (ctx.guild_id, role.id))
-        ])
+        db.multi_void_sql([
+            ("DELETE FROM RoleChannel WHERE GuildID=%s AND RoleID=%s", (ctx.guild_id, role.id)),
+            ("DELETE FROM ReactMessages WHERE GuildID=%s AND RoleID=%s", (ctx.guild_id, role.id))])
         await ctx.respond(f"All role assign behaviours have been cleared for {role.name}")
 
     @discord.slash_command(name="clear_role_setting",
@@ -79,7 +82,7 @@ class Admin(commands.Cog):
     @discord.commands.default_permissions(manage_guild=True)
     async def delete_react_entry(self, ctx: discord.ApplicationContext) -> None:
         logger.info("Dropping react entries", guild_id=ctx.guild_id)
-        db.multi_void_SQL([
+        db.multi_void_sql([
             ("DELETE FROM ReactMessages WHERE GuildID=%s", (ctx.guild_id,)),
             ("DELETE FROM RoleChannel WHERE GuildID=%s", (ctx.guild_id,))
         ])
@@ -95,7 +98,9 @@ class Admin(commands.Cog):
         channel: discord.TextChannel,
         role: discord.Role
     ) -> None:
-        db.single_void_SQL("INSERT INTO RoleChannel VALUES (%s, %s, %s, TRUE)", (ctx.guild_id, role.id, channel.id))
+        db.single_void_SQL(
+            "INSERT INTO RoleChannel VALUES (%s, %s, %s, TRUE)",
+            (ctx.guild_id, role.id, channel.id))
         await ctx.respond(f"Role channel was set to {channel.mention}")
 
     @discord.slash_command(name='set_remove_role_channel',
@@ -108,13 +113,16 @@ class Admin(commands.Cog):
         channel: discord.TextChannel,
         role: discord.Role
     ) -> None:
-        db.single_void_SQL("INSERT INTO RoleChannel VALUES (%s, %s, %s, FALSE)", (ctx.guild_id, role.id, channel.id))
+        db.single_void_SQL(
+            "INSERT INTO RoleChannel VALUES (%s, %s, %s, FALSE)",
+            (ctx.guild_id, role.id, channel.id))
         await ctx.respond(f"Role remove channel was set to {channel.mention}")
 
     @discord.commands.slash_command(
         name="set_chain_message",
         description="allows Nix to follow up with custom messages whenever a user send a message")
-    @discord.commands.option("message", type=str, description="The message text that will be sent as a follow up. " +
+    @discord.commands.option("message", type=str,
+                             description="The message text that will be sent as a follow up. " +
                              "Write <<user>> to ping the user")
     @discord.commands.option("response_channel", type=discord.TextChannel,
                              description="The channel where Nix sends its follow up")
@@ -137,12 +145,15 @@ class Admin(commands.Cog):
             )
             await ctx.respond(f"You set a chain_message for the channel {response_channel}")
         except db.KeyViolation:
-            await ctx.respond(f"You already set this as a message for this channel {Emotes.CONFUSED}")
+            await ctx.respond(
+                f"You already set this as a message for this channel {Emotes.CONFUSED}"
+            )
 
-    @commands.slash_command(name="clear_chain_messages", description="Clears all chain message behaviours")
+    @commands.slash_command(name="clear_chain_messages",
+                            description="Clears all chain message behaviours")
     @discord.commands.default_permissions(manage_guild=True)
     async def clear_chain_message(self, ctx: discord.ApplicationContext) -> None:
-        db.multi_void_SQL([
+        db.multi_void_sql([
             ("DELETE FROM ChainedUsers WHERE GuildID=%s", (ctx.guild_id,)),
             ("DELETE FROM MessageChain WHERE GuildID=%s", (ctx.guild_id,))
         ])
@@ -159,15 +170,16 @@ class Admin(commands.Cog):
             logger.error("Bot is offline", channel_id=msg.channel.id)
             return
         if msg.author.id != self.bot.user.id:
-            values = db.single_SQL("SELECT WatchedChannelID FROM MessageChain WHERE GuildID=%s", (msg.guild.id,))
+            values = db.single_sql(
+                "SELECT WatchedChannelID FROM MessageChain WHERE GuildID=%s", (msg.guild.id,))
             if values is not None:
                 check_vals = [val[0] for val in values]
                 if msg.channel.id in check_vals or -1 in check_vals:
                     try:
                         db.single_void_SQL(
                             "INSERT INTO ChainedUsers VALUES (%s, %s, %s)",
-                            (msg.guild.id, msg.author.id, msg.channel.id if msg.channel.id in check_vals else -1)
-                        )
+                            (msg.guild.id, msg.author.id, msg.channel.id
+                             if msg.channel.id in check_vals else -1))
                         await self.send_chained_message(msg.guild, msg.author)
                     except db.KeyViolation:
                         logger.info(
@@ -190,7 +202,8 @@ class Admin(commands.Cog):
             logger.info("Author is not member (likely: user not in guild)")
             return
         if msg.author.id != self.bot.user.id:
-            vals = db.single_SQL("SELECT RoleID, ToAdd FROM RoleChannel WHERE ChannelID=%s", (msg.channel.id,))
+            vals = db.single_sql(
+                "SELECT RoleID, ToAdd FROM RoleChannel WHERE ChannelID=%s", (msg.channel.id,))
             for (role_id, add_role) in vals:
                 role = msg.guild.get_role(role_id)
                 if role:
@@ -212,7 +225,8 @@ class Admin(commands.Cog):
             logger.info("reaction event has no member (likely: user not in guild)")
             return
         logger.debug(f"Message ID on reaction: {event.message_id}")
-        vals = db.single_SQL("SELECT Emoji, RoleID FROM ReactMessages WHERE MessageID=%s", (event.message_id,))
+        vals = db.single_sql(
+            "SELECT Emoji, RoleID FROM ReactMessages WHERE MessageID=%s", (event.message_id,))
         logger.debug(f"SQL values: {vals}")
         for (emoji, role_id) in vals:
             if Emoji(emoji).to_partial_emoji() == event.emoji:
@@ -226,9 +240,11 @@ class Admin(commands.Cog):
     @commands.Cog.listener('on_raw_reaction_remove')
     async def unassign_react_role(self, event: discord.RawReactionActionEvent) -> None:
         if event.guild_id is None:
-            logger.info("unassign_react_role detected outside of guild", channel_id=event.channel_id)
+            logger.info("unassign_react_role detected outside of guild",
+                        channel_id=event.channel_id)
             return
-        vals = db.single_SQL("SELECT Emoji, RoleID FROM ReactMessages WHERE MessageID=%s", (event.message_id,))
+        vals = db.single_sql(
+            "SELECT Emoji, RoleID FROM ReactMessages WHERE MessageID=%s", (event.message_id,))
         for (emoji, role_id) in vals:
             if Emoji(emoji).to_partial_emoji() == event.emoji:
                 logger.debug("removing role")
@@ -241,19 +257,24 @@ class Admin(commands.Cog):
                     logger.error("Couldnt get role for react role unassign")
 
     @staticmethod
-    async def send_chained_message(guild: discord.Guild, user: discord.User | discord.Member) -> None:
-        vals = db.single_SQL("SELECT ResponseChannelID, Message FROM MessageChain WHERE GuildID=%s", (guild.id,))
+    async def send_chained_message(
+        guild: discord.Guild,
+        user: discord.User | discord.Member
+    ) -> None:
+        vals = db.single_sql(
+            "SELECT ResponseChannelID, Message FROM MessageChain WHERE GuildID=%s", (guild.id,))
         for (response_channel_id, message) in vals:
             msg = message.replace("<<user>>", user.mention)
             try:
                 channel = guild.get_channel(response_channel_id)
                 if not isinstance(channel, discord.abc.Messageable):
-                    logger.error("Chained message channel not sendable", channel_id=guild.id, guild_id=guild.id)
+                    logger.error("Chained message channel not sendable",
+                                 channel_id=guild.id, guild_id=guild.id)
                     continue
                 await channel.send(msg)
             except discord.errors.Forbidden:
-                logger.info("Permission failure for chain_message", guild_id=guild.id, channel_id=response_channel_id)
-                pass
+                logger.info("Permission failure for chain_message",
+                            guild_id=guild.id, channel_id=response_channel_id)
 
 
 def setup(bot: discord.Bot) -> None:
